pyproject.toml
[project]
name = "expense-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "coreapi>=2.3.3",
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "google-genai>=1.68.0",
    "google-generativeai>=0.8.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
